feat: increase rate and automate current month updates
This will improve jobs database closer
to real-time updates to the hiring thread without
the need for human verfication to ship to production.

A final human validation is required on the final
markdown output. This is a temporary failsafe to
prevent polluting the `main` branch. It will be
automated in future releases as the jobs.db works.